Avoid unnecessary explicit cast
...which would only hide unexpected conversion failure, should it ever be
possible to happen, from tools like -fsanitize=integer

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/86977
Tested-by: Jenkins

+++ sfx2/source/bastyp/mieclip.cxx
@@ -79,7 +79,7 @@ SvStream* MSE40HTMLClipFormatObj::IsValid( SvStream& rStream )
 
     if (nFragStart > 0 && nFragEnd > 0 && nFragEnd > nFragStart)
     {
-        size_t nSize = static_cast<size_t>(nFragEnd - nFragStart + 1);
+        size_t nSize = nFragEnd - nFragStart + 1;
         if (nSize < 0x10000L)
         {
             rStream.Seek(nFragStart);
